Don't redefine macros with OCaml 4.12
Macros `Val_none` and `Some_val` are part of the standard library
starting with OCaml 4.12.

In some environments, redefining a macro triggers a compiler warning.
When warnings are treated as errors, the build fails.

Issue discussing the introduction of the macros, the problem, and the
fix: https://github.com/ocaml/ocaml/issues/5154

Commit introducing the macros in the standard library:
https://github.com/ocaml/ocaml/commit/973eeb1867f41ef2669ad4cbfa9c2e76ddef1749